Move JS spec files into spec folder
We pinned all JS files under app/javascript into importmap
and do not want to load test files.
